Remove 'skip' from include_tasks
The 'skip' attribute of a 'include_tasks' module was set to 'true' to
allow not including any task file if there were no matching file found.

However this is not the case (there's always something to include,
otherwise the 'include_tasks' module wouldn't be used) and it can easily
hide errors.

So now at least one of the files must be included or an error will be
thrown.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>