config.collect: Continue in the absence of test-info file
In those situations where environment was not setup correctly we fail to
run any tests. This would mean that test-info.yml file is not linked and
subsequent task to copy it fails. Therefore make sure that we collect
rest of the logs and configuration by ignoring the errors while copying
the test-info file.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>