test.sit-test-cases: Redirect test output to a file
Instead of cluttering the jenkins job console output with detailed test
results we could redirect everything to a separate file and make it
available as artifacts for each job. Simplest way to achieve this
redirection is to make use of '&>'[1] to include both STDOUT and STDERR.
Location of such an output file is chosen to be under /var/log so that
it also gets collected as part of statedump process.

[1] https://www.gnu.org/software/bash/manual/html_node/Redirections.html(3.6.4)

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>