local.defaults: Change default backend to XFS
It has been a while since XFS was added to the list of supported
backends and is known to successfully complete full test runs. On
the other side GlusterFS is going through tough times in getting
necessary fixes for long pending failures. Let's flip the switch
and change the default backend from GlusterFS to XFS.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>