Add support for 'gpfs.vfs' backend variant
Now it's possible to use the 'vfs' variant of the GPFS backend. With it,
the gpfs VFS module is automatically configured to be used by Samba.

The GPFS VFS module dynamically loads libgpfs.so library, which requires
access to some resources that are not granted by default by SElinux. The
Storage Scale product solves this by running smbd process unconfined, so
the same solution has been applied here.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>