Fix share permissions for gpfs.scale
When the 'scale' variant of gpfs backend is used, the regular samba and
ctdb roles are not installed. The permissions to the root directory of
the share were defined in these roles.

This patch moves the configuration of the permissions (for all backends)
to the new_volume.yml file that is always used to create a new share,
even for the 'scale' variant.

Fixes: #92
Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>