sit-xfs: do not recreate fs
At the moment the playbook forces the creation of the xfs fs. This means
that when re-running the ansible playbook, we attempt to recreate a
filesystem which has already been mounted. This fails causing it to
fail.

By setting to false, we do not force a recreation of the filesystem
which prevents this operation being idempotent.

Signed-off-by: Sachin Prabhu <sp@spui.uk>