hosts.update: Remove system installed libldb
With recent changes to how libldb upstream sources are distributed,
we are observing conflict for the bundled libldb files when there is
already a newer version of libldb present in the system. Since we do
not rely on system installed libldb at all, remove it as first step to
avoid conflicts.

ref: https://github.com/samba-in-kubernetes/samba-build/pull/66

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>