test-info: Include share information from individual files
With increasing number of variations to how shares for different
backends are populated, the template for test-info.yml has become
convoluted. Moreover it lost its generic nature due to additional
checks on various backend specific options while forming the shares
section. Thus we split the shares section in to separate backend
specific templates and include it in the main test-info.yml.

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>